requirements.txt
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.0
google-cloud-pubsub==2.21.1
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
flask==3.0.0
anthropic==0.40.0
supabase==2.0.0
authlib==1.3.0
Flask-Session==0.6.0
